tests/submission/Infraware/Forms/contents/Forms/button_formenctype.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>button_formEnctype</h3>
	</p>
	
	<hr>

	<div id="log"></div>
	
	<form method="post"
		  enctype=""
		  action=""
		  name="input_form">
	<p><button id='button_id'>button</button></p>
	</form>

	<script>
		
		var button = document.getElementById("button_id");
		button.formEnctype = "application/x-www-form-urlencoded";

		if (typeof(button.formEnctype) == "string") {
			test(function() {
				assert_equals(button.formEnctype, "application/x-www-form-urlencoded", "formEnctype attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formEnctype attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
